vmshutdown/requirements.txt
azure-mgmt-compute
azure-mgmt-resource
azure-common
datetime